requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
razorpay==1.4.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
